pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vecapp-ai"
version = "1.0.0"
description = "AI service for analyzing and summarizing follow-up interactions with visitors"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Determinas Technologies", email = "info@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
]
dependencies = [
    # Web Framework
    "fastapi==0.109.0",
    "uvicorn[standard]==0.34.2",
    "python-dotenv==1.0.0",
    "requests==2.32.3",
    "click==8.1.7",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "h11==0.14.0",
    "pydantic==2.7.3",
    "pydantic-core==2.18.4",
    "pydantic-settings==2.2.1",
    "starlette==0.35.1",
    "slowapi",
    "python-multipart==0.0.9",
    
    # Database
    "asyncpg==0.29.0",
    "sqlalchemy>=2.0.12,<3.0.0",
    "alembic==1.13.1",
    "databases==0.9.0",
    "psycopg2-binary==2.9.9",
    
    # Messaging
    "kafka-python==2.0.2",
    "aiokafka==0.10.0",
    "redis>=4.5.5,<5.0.0",
    "boto3==1.34.162",
    
    # AI/ML
    "openai==1.12.0",
    "numpy==1.26.4",
    "scikit-learn==1.4.0",
    "pandas==2.2.0",
    
    # Authentication & Security
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "bcrypt==4.1.2",
    
    # LangChain & Ecosystem
    "langchain==0.1.0",
    "langchain-openai==0.0.5",
    "langchain-community==0.0.13",
    "langchain-core==0.1.18",
    "langsmith==0.0.83",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/vecapp-ai"
"Bug Tracker" = "https://github.com/yourusername/vecapp-ai/issues"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.alembic]
script_location = "app/database/migrations"